Move Havoc to talents folder because it is a talent
