+++ src/analysis/retail/warlock/destruction/CombatLogParser.ts
@@ -12,7 +12,7 @@ import AlwaysBeCasting from './modules/features/AlwaysBeCasting';
 import Backdraft from './modules/talents/Backdraft';
 import Checklist from './modules/features/Checklist/Module';
 import CooldownThroughputTracker from './modules/features/CooldownThroughputTracker';
-import Havoc from './modules/features/Havoc';
+import Havoc from './modules/talents/Havoc';
 import ImmolateUptime from './modules/features/ImmolateUptime';
 import RainOfFire from './modules/features/RainOfFire';
 import SoulShardDetails from './modules/soulshards/SoulShardDetails';
